refactor: sse 위치 변경
