+++ tracky-web/src/main/java/kernel360/trackyweb/rent/application/RentService.java
@@ -15,8 +15,8 @@ import kernel360.trackycore.core.common.entity.CarEntity;
 import kernel360.trackycore.core.common.entity.RentEntity;
 import kernel360.trackycore.core.common.provider.CarProvider;
 import kernel360.trackycore.core.common.provider.RentProvider;
-import kernel360.trackycore.core.common.sse.GlobalSseEvent;
-import kernel360.trackycore.core.common.sse.SseEvent;
+import kernel360.trackyweb.common.sse.GlobalSseEvent;
+import kernel360.trackyweb.common.sse.SseEvent;
 import kernel360.trackyweb.rent.application.dto.request.RentRequest;
 import kernel360.trackyweb.rent.application.dto.response.RentResponse;
 import kernel360.trackyweb.rent.domain.provider.RentDomainProvider;
